requirements/base.txt
Django==1.11.7
django-allauth==0.34.0
django-braces==1.11.0
django-constance[database]==2.0.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-extensions==1.9.7
django-model-utils==3.0.0
django-solo==1.1.2
django-silk==1.0.0
djangorestframework==3.7.3
django-rest-auth==0.9.2
django-phonenumber-field==1.3.0
django-rest-swagger==2.1.2
-e git+https://github.com/steelkiwi/drf-skd-tools@0.1.2#egg=drf_skd_tools
googlemaps==2.5.1
Faker==0.7.3
mixer==5.6.6
django-prettyjson==0.3.0
django-admin-sortable2==0.6.19
django-scheduler==0.8.8
structlog==18.1.0
django-cities-light==3.3.0
django-bower==5.2.0
boto3==1.6.3
django-storages==1.6.5
PyJWT==1.6.0
MonthDelta==0.9.1
django-polymorphic==2.0.2
django-modeltranslation==0.12.2
django-filter==1.1.0
django-ckeditor==5.4.0
sorl-thumbnail==12.4.1
reverse-geocoder==1.5.1

Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
raven==6.3.0
celery==4.1.0
redis==2.10.6
